Estimate Causal Effects

This is a generic function that estimates causal effects for different object types. It automatically dispatches the appropriate method based on the class of `object`.

Usage

est_effect(object, Y, ...)

Arguments

object

An object to calculate effects.

Y

Numeric vector of observed outcomes.

...

Additional arguments (ignored).

Value

Returns values of effects.

Details

This function uses S3 method dispatching to call the appropriate method based on the object type. It is designed for estimating causal effects in settings with continuous or binary outcomes. For survival outcomes, users should apply appropriate survival analysis models, such as a weighted Cox model or other time-to-event estimation methods. For `lbc_net` objects, see est_effect.lbc_net for more details. Additional object types may be supported in the future.